requirements.txt
# **FastAPI and Web Server**
fastapi
uvicorn[standard]
# **AI Adapter Core Dependencies**
pydantic
pyyaml
python-dotenv
openai
google-genai
ollama
# **Database (to be added in the future)
sqlalchemy
alembic
psycopg2-binary
